Try using boost multi_index, failed
Try using boost multi_index to list enum map by enum value.
However, using enum value as index causes `find()` searching
for value instead of string. The error message indicates that
the code tries to `find(std::string)` which is not comparable
with the enum value.

That is, the sorting and searching will be bound to the same
index, thus fails to achive our goal.